채팅창 기본정보 Dialog 완성

+++ app/src/main/java/net/simplifiedcoding/firebaseauthtutorial/adapter/SummonerInfo.kt
@@ -1,20 +1,20 @@
 package net.simplifiedcoding.firebaseauthtutorial.adapter
 
 data class SummonerInfo(
-    val nickname : String,
-    val level : Integer,
-    val typeB : String,
-    val typeM : String,
-    val typeJ : String,
-    val typeT : String,
-    val winrateB : Double,
-    val winrateM : Double,
-    val winrateJ : Double,
-    val winrateT : Double,
-    val KDA_B : Double,
-    val KDA_M : Double,
-    val KDA_J : Double,
-    val KDA_T : Double,
-    val KDA_Total : Double,
-    val winrateTotal : Double
+    val nickname: String,
+    val level: Long,
+    val typeB: String,
+    val typeM: String,
+    val typeJ: String,
+    val typeT: String,
+    val winrateB: Double,
+    val winrateM: Double,
+    val winrateJ: Double,
+    val winrateT: Double,
+    val KDA_B: Double,
+    val KDA_M: Double,
+    val KDA_J: Double,
+    val KDA_T: Double,
+    val KDA_Total: Double,
+    val winrateTotal: Double
 )
